+++ src/main/java/Interfaces/ISell.java
@@ -1,5 +1,5 @@
 package Interfaces;
 
 public interface ISell {
-    double calculateMarkup();
+    int calculateMarkup();
 }
